refactor(jaspr): tidy SizedBox child and geometry rect construction
- SizedBox: express the optional child as a switch rather than a collection-if.
- Jaspr geometry rectOf: DOMRect accessors already return double, so drop the
  redundant toDouble() casts.

No behavior change; full check.sh green (incl. headless-Chrome geometry).

+++ packages/a2ui_craft_jaspr/test/flex_geometry_test.dart
@@ -50,10 +50,10 @@ class _JasprGeometryTester implements CraftGeometryTester {
         _tester.findNode<web.Element>(find.byKey(ValueKey<String>(key)))!;
     final web.DOMRect r = el.getBoundingClientRect();
     return CraftRect(
-      left: r.left.toDouble(),
-      top: r.top.toDouble(),
-      width: r.width.toDouble(),
-      height: r.height.toDouble(),
+      left: r.left,
+      top: r.top,
+      width: r.width,
+      height: r.height,
     );
   }
 }
